Add sticky position to header in CSS file
The CSS file has been modified to make the header sticky. Some spaces were also added to improve the treatment of indentation in the file, making the code more readable and maintainable. The change provides a static navigation bar that is fixed at the top of the viewport when users scroll.